requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.5
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-extensions==3.1.5
django-import-export==2.8.0
django-mathfilters==1.0.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
executing==0.8.3
extensions==0.4
google==3.0.0
gunicorn==20.1.0
idna==3.3
ipython==8.4.0
jdcal==1.4.1
jedi==0.18.1
MarkupPy==1.14
matplotlib-inline==0.1.3
odfpy==1.4.1
openpyxl==3.0.10
paho-mqtt==1.6.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==4.21.2
#psycopg2==2.9.3
pure-eval==0.2.2
pydotplus==2.0.2
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
tablib==3.2.1
traitlets==5.3.0
twilio==7.12.0
tzdata==2022.1
urllib3==1.26.9
validate-email==1.3
wcwidth==0.2.5
Werkzeug==2.1.2
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
